+++ commandry-core/src/main/java/de/eldoria/commandry/parser/ParserManager.java
@@ -1,13 +1,6 @@
 package de.eldoria.commandry.parser;
 
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * This class manages multiple parser instances.
- */
-public class ParserManager {
-    private final Map<Class<?>, Parser<?>> parsers = new HashMap<>();
+public interface ParserManager {
 
     /**
      * Registers a parser for the given type. Calling this method may replace
@@ -17,9 +10,7 @@ public class ParserManager {
      * @param clazz  the type class to register the parser for.
      * @param <T>    the type which can be parsed by the parser.
      */
-    public <T> void registerParser(Parser<T> parser, Class<T> clazz) {
-        parsers.put(clazz, parser);
-    }
+    <T> void registerParser(Parser<T> parser, Class<T> clazz);
 
     /**
      * Returns whether there is an existing parser for the given type.
@@ -27,9 +18,7 @@ public class ParserManager {
      * @param clazz the type class to look up.
      * @return {@code true} if a parser for the given type exists, {@code false} otherwise.
      */
-    public boolean hasParserFor(Class<?> clazz) {
-        return parsers.containsKey(clazz);
-    }
+    boolean hasParserFor(Class<?> clazz);
 
     /**
      * Parses an input string to the requested type. If no parser is found,
@@ -40,12 +29,5 @@ public class ParserManager {
      * @param <T>    the type to return.
      * @return the parsed object of the requested type.
      */
-    public <T> T parse(String input, Class<T> target) {
-        Parser<?> parser = parsers.get(target);
-        if (parser == null) {
-            // TODO
-            return null;
-        }
-        return (T) parser.parse(input);
-    }
+    <T> T parse(String input, Class<T> target);
 }
